move ResourceViewUtils to base-tool module
