+++ ui/base-tool/build.gradle
@@ -1,4 +1,5 @@
 dependencies {
+    implementation project(':library:base')
     implementation project(':library:xml-model')
 
     api project(':ui:base')
