+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/ImageViewHolder.java
@@ -6,6 +6,7 @@ import android.support.annotation.UiThread;
 import android.view.ViewGroup;
 
 import org.ccci.gto.android.common.picasso.view.PicassoImageView;
+import org.cru.godtools.base.tool.model.view.ResourceViewUtils;
 import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
 import org.cru.godtools.xml.model.Image;
