+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/ManifestViewUtils.java
@@ -5,6 +5,7 @@ import android.graphics.Typeface;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import org.cru.godtools.base.tool.model.view.ResourceViewUtils;
 import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.base.ui.util.LocaleTypefaceUtils;
 import org.cru.godtools.xml.model.Manifest;
@@ -17,7 +18,7 @@ public class ManifestViewUtils {
 
     public static void bindBackgroundImage(@Nullable final Manifest manifest,
                                            @NonNull final ScaledPicassoImageView view) {
-        ResourceViewUtils.bindBackgroundImage(view, Manifest.getBackgroundImageResource(manifest),
+        ResourceViewUtils.bindBackgroundImage(Manifest.getBackgroundImageResource(manifest), view,
                                               Manifest.getBackgroundImageScaleType(manifest),
                                               Manifest.getBackgroundImageGravity(manifest));
     }
