+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/PageViewHolder.java
@@ -15,6 +15,7 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
 
 import org.cru.godtools.base.model.Event;
+import org.cru.godtools.base.tool.model.view.ResourceViewUtils;
 import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.tract.R2;
 import org.cru.godtools.tract.widget.PageContentLayout;
@@ -217,7 +218,7 @@ public class PageViewHolder extends ParentViewHolder<Page>
 
     private void bindPage() {
         mPageView.setBackgroundColor(Page.getBackgroundColor(mModel));
-        ResourceViewUtils.bindBackgroundImage(mBackgroundImage, Page.getBackgroundImageResource(mModel),
+        ResourceViewUtils.bindBackgroundImage(Page.getBackgroundImageResource(mModel), mBackgroundImage,
                                               Page.getBackgroundImageScaleType(mModel),
                                               Page.getBackgroundImageGravity(mModel));
     }
